Move DFS to package graph; Add computeIfAbsent; remove unnecessary braces in loop and condition.
fix errors in test
