+++ src/test/java/com/github/olegik1719/trello/graph/DFSTest.java
@@ -1,4 +1,4 @@
-package com.github.olegik1719.trello;
+package com.github.olegik1719.trello.graph;
 
 import org.junit.Before;
 import org.junit.Test;
@@ -43,8 +43,8 @@ public class DFSTest {
         assertFalse(unaccessible.isAccessible(cities[0],cities[7]));
     }
 
-    private void addEdges(DFS dfs, int[] data){
-    for (int i=1; i<data.size; i++)
-        dfs.addEdge(cities[data[0]], cities[data[i]])
+    private void addEdges(DFS dfs, int... data){
+    for (int i=1; i<data.length; i++)
+        dfs.addEdge(cities[data[0]], cities[data[i]]);
     }
 }
